==> R/01_load.qmd <==
---
title: "01_load"
format: html
execute:
  standalone: true
---

```{r}
#| message: false

library("tidyverse")
```

## Making sure the directories exist

```{r}
#| message: false
#| warning: false
raw_dir <- "../data/_raw/"

if( !dir.exists(raw_dir) ){
  dir.create(path = raw_dir)
}
```

Please see the README to for the data retrieval and correct placement.

```{r}
#| message: false
#| warning: false
raw_data <- read_csv("../data/_raw/SAA_Biospecimen_Analysis_Results_01Nov2025.csv")
```

## Saving data-frame

```{r}
write_tsv(raw_data, "../data/01_dat_load.tsv")
```
