==> R/02_clean.qmd <==
---
title: "02_clean"
format: html
---

```{r}
#| message: false

library(tidyverse)
```

## Loading the raw data

```{r}
#| message: false

raw_data <- read_tsv("../data/01_dat_load.tsv")
```

## Clean the data

### Selecting only PD patients:

SWEDD is an outdated ID, mentioned in the PPMI data guide, and we will not be using the data from these patients.

```{r}
tidy_data <- raw_data |>
  filter(COHORT != "SWEDD")
```

### Deleting columns

Some columns in the data-set are not part of the chosen analytic points. The columns chosen to omit is at the cause are either:

-   Irrelevant to the analysis such as the name of the institution or the responsible practitioner.

-   Contain the same value for all observation such as the TYPE (which is always a cerebrospinal fluid sample) and SAAMethod (results always made on the basis of a SAA).

-   Only have a value in observed in one of the projects: SAA_Type, TSmax, T50.

-   Lack meassurements for most observations like SampleVolume.

```{r}
# Neglecting the unused columns:
tidy_data <- tidy_data |>
  select(-"COHORT",
         -"PI_NAME",
         -"PI_INSTITUTION",
         -"TYPE",
         -"SAAMethod",
         -"SAA_Type",
         -starts_with("TSmax"),
         -starts_with("SLOPE"),
         -starts_with("T50"),
         -starts_with("Sample"))

```

### Missing value interpretation

One person (patient_number = 41184) has the value "U01" in the clinical event column. As no information is found explaining the value, the patient is removed from the data.

```{r}
tidy_data <- tidy_data |>
  filter(CLINICAL_EVENT != "U01")
```

### Elongating the data

Originally all repetitions of Fmax, TTT and AUC had its own column. To make the data more comprehensible all repetitions are stored in one column for each measurement. The data will then contain only one column of each measurement (Fmax/AUC/TTT), a column with the duration (either 24 or 150 hours) and a column defining the repetition number (1-3) of the measurement.

```{r}
tidy_data <- tidy_data |> 
  pivot_longer(
    cols = matches("Fmax|TTT|AUC"), 
    names_to = c("measurement", "duration", "rep"), 
    names_pattern = ("(Fmax|TTT|AUC)_(\\d+)h_Rep(\\d+)"),
    values_to = "value",
    values_drop_na = TRUE
  ) |> 
  
  pivot_wider(
    names_from = measurement,
    values_from = value
  )
```

### Reducing the "Instrument" columns

When looking at the data it seemed that the instrument used for one observation for all three repetitions was the same. We found that it was the case for all observations when checking:

```{r}
all(
  tidy_data$InstrumentRep1 == tidy_data$InstrumentRep2 & tidy_data$InstrumentRep2 == tidy_data$InstrumentRep3
  )

```

To reduce the number of columns, the instrument is stored in a single column.

```{r}
tidy_data <- tidy_data |> 
  select(!InstrumentRep2 & !InstrumentRep3) |> 
  rename(instrument = InstrumentRep1)
```

### Renaming the columns to preferred naming convention

```{r}
tidy_data <- tidy_data |>
  rename(
    patient_number = PATNO,
    sex = SEX,
    project = PROJECTID,
    clinical_event = CLINICAL_EVENT,
    saa_result = SAA_Status,
    rundate = RUNDATE,
    fmax = Fmax,
    ttt = TTT,
    auc = AUC)
tidy_data |> 
  slice_sample(n = 10)
```

### Changing the column-types

```{r}
source("09_proj_func.R")
tidy_data <- char_type_col(tidy_data)
```

### Relocating columns

The columns containing important information about an observation was moved to the front.

```{r}
tidy_data <- tidy_data |>
	relocate(project)

tidy_data <- tidy_data |>
	relocate(duration:auc, .after = saa_result)
```

### Cluster the patients' observations

Some patients have multiple observations and uses the patient_number as an identifier. One patients observations are all gathered together by arranging the patient_number and individually put in order depending on their visit.

```{r}
tidy_data <- tidy_data |>
  arrange(patient_number, rundate)
```

## Save the cleaned data

```{r}
write_tsv(tidy_data, "../data/02_dat_clean.tsv")
```
