==> R/06_analysis_2.qmd <==
---
title: "06_analysis_2"
format: html
---

## Loading libraries and loading the data

```{r}
#| message: false
library(tidyverse)

aug_data <- read_tsv("../data/03_aug_data.tsv")

#making sure that the columns is the right type
source("09_proj_func.R")
aug_data <- char_type_col(aug_data)
```

## Defining colors and theme

```{r}
my_cols <- c("#da9283" ,"#7A1F15","#94BFBE","#627634","#C03221","#545E75",  
              "#3C3C68","#474B24")

theme_custom <- theme_minimal(base_size = 14) +
  
  theme(
    plot.title = element_text(face = "bold", 
                              size = 13, 
                              margin = margin(b = 7)
                              ),
    
    plot.subtitle = element_text(size = 10, 
                                 margin = margin(b = 12)
                                 ),
    
    plot.caption = element_text(size = 8, 
                                margin = margin(t = 20)
                                ),
      
    axis.title.x = element_text(size = 10, 
                                margin = margin(t = 10)
                                ),
    
    axis.title.y = element_text(size = 10, 
                                margin = margin(r = 10)
                                ),

    axis.text.x = element_text(size = 10, 
                               margin = margin(t = 5)
                               ),
    
    axis.text.y = element_text(size = 10, 
                               margin = margin(r = 5)
                               ),
    
    legend.position = "none",
    )
```

## Analysis of Fmax and SAA results

Seed Amplification Assays (SAA) detect α-synuclein aggregates in cerebrospinal fluid (CSF) and monitors fluorescence intensity over time. The point at which fluorescence reaches its maximum (Fmax) reflects the level of α-synuclein aggregation and is therefore a potential diagnostic indicator for Parkinson’s disease (PD).

The two projects perform the assays under different variables:

-   **Project 155**: 150-hour assay, fluorescence measured every 29 minutes

-   **Project 237**: 24-hour assay, fluorescence measured every 14 minutes

Variations in assay duration, cycle length, and reaction mixture composition mean that fluorescence values (Fmax) are not directly comparable across projects.

Because assay duration, cycle frequency, and reaction mixture differ between projects, fluorescence values (Fmax) cannot be directly compared without looking at protocol-specific characteristics.

We will look at the separation and overlap in fmax, when determining SAA result.

## SAA result distribution

We will look at how the ratio of negative, positive and inconclusive is

```{r}
saa_result_dist <- aug_data |>
  ggplot(mapping = aes(x = project)) + 
  geom_bar(mapping = aes(fill = saa_result),
           position = "fill") + 
  
  scale_fill_manual(values = c("#da9283","#7A1F15","#94BFBE"),
                    name = "SAA result",
                    guide = guide_legend(nrow = 1)) +
  
  theme_custom + 
  theme(legend.position = "bottom") + 
  labs(title = "SAA result distribution",
       subtitle = "SAA result ratio from the two projects",
       x = "Project",
       caption = "Source: PPMI (Parkinson's Progression Markers Initiative)") +
    
    scale_y_continuous(name = "Percentage", 
                       labels = scales::label_percent(accuracy = 1))

ggsave(filename = "../results/06_saa_result_dist.png",
       width = 10,  
       height = 5)   

saa_result_dist
```

(very hard to figure out how to do the percent labels) **stacked plot with percent labels:**

```{r}
saa_result_percent <- aug_data |> 
  group_by(project) |> 
  count(saa_result) |> 
  mutate(pct = prop.table(n)) |> 
  
  ggplot(aes(x = project, 
             y = pct, 
             fill = saa_result, 
             label = scales::percent(pct))) + 
  
  geom_col(position = position_stack(reverse = TRUE)) + 
  
  geom_label(position = position_stack(reverse = TRUE),
            fill = "white",
            alpha = 0.8, 
            label.size = 0,
            size = 3) + 
  
  scale_y_continuous(labels = scales::percent) +
  
  scale_fill_manual(values = c("#da9283","#7A1F15","#94BFBE"),
                    name = "SAA result",
                    guide = guide_legend(nrow = 1)) +
  theme_custom + 
  theme(legend.position = "bottom") + 
  labs(title = "SAA result distribution",
       subtitle = "SAA result percentages from project 155 & 237",
       y = "Percent",
       x = "Project",
       caption = "Source: PPMI (Parkinson's Progression Markers Initiative)")
  
ggsave(filename = "../results/06_saa_result_percent.png",
       width = 10,  
       height = 5)   

saa_result_percent
```

It seems to be very similar - not a very big difference in the ratio. Trying to see the separation:

## Analysis of Fmax_mean as alternative test parameter

In this section we will test the variable Fmax_mean as a valid predictor of test result status. Fmax_mean is a mean calculated from the three samples collected in every assay.

## Test distinction by way of Fmax_mean

First we select the 150h test group from the dataset, and then we plot the Fmax_mean grouped by their result status(positive/negative/inconclusive) and do a visual inspection and evaluate if the groups are distinguishable.

```{r}
fmax_mean_results_150h <- aug_data |> 
  filter(project == "155") |> 
  ggplot(aes(
    x = saa_result,
    y = fmax_mean,
    fill = saa_result)
    ) +
  
  geom_boxplot(outlier.alpha = 0.3) +
  
  scale_color_manual(values = my_cols) +
  scale_fill_manual(values = my_cols) +
  
  labs(
    title = "150h Metrics by SAA Status",
    y = "Fmax mean",
    x = NULL,
    caption = "Source: PPMI (Parkinson's Progression Markers Initiative)"
    ) +
  theme_custom
```

```{r}
fmax_mean_results_150h

ggsave(filename = "../results/06_fmax_mean_results_150h.png",        
       width = 10,
       height = 5)  
```

From the plot we see that that the positive and negative group is distinguishable so we conduct a t-test to evaluate if we can predict what group a measurement would belong to with confidence.

First we filter for positive and negative and drop NA values.

```{r}
filtered_155 <- aug_data |> 
  filter(saa_result %in% c("Positive", 
                           "Negative"), 
         project == "155") |>   
  
  drop_na(fmax_mean) |>
  
  select(patient_number, 
         saa_result, 
         fmax_mean)
```

Then we conduct the T-test

```{r}
t.test(fmax_mean ~ saa_result, 
       data = filtered_155)
```

Here the T-test shows a highly statistically significant difference between the groups.

## We repeat the steps for the 24h test group

```{r}
fmax_mean_results_24h <- aug_data |>
  filter(project == "237") |>
  ggplot(aes(
    x = saa_result,
    y = fmax_mean,
    fill = saa_result)
    ) +
  
  geom_boxplot(outlier.alpha = 0.3) +
  scale_color_manual(values = my_cols) +
  scale_fill_manual(values = my_cols) +
  
  labs(title = "24h Metrics by SAA Status",
       y = "Fmax mean",
       x = NULL,
       caption = "Source: PPMI (Parkinson's Progression Markers Initiative)") +
  
  theme_custom
```

```{r}
fmax_mean_results_24h

ggsave(filename = "../results/06_fmax_mean_results_24h.png",
       width = 10,
       height = 5)  
```

```{r}
filtered_237 <- aug_data |>
  filter(saa_result %in% c("Positive", 
                           "Negative"),
         project == "237") |> 
  
  drop_na(fmax_mean) |>
  
  select(patient_number, 
         saa_result, 
         fmax_mean)  

filtered_237 |> count(saa_result) 
```

```{r}
t.test(fmax_mean ~ saa_result, 
       data = filtered_237)
```

Here we also show the ability to visually distinguish between the groups and we show a highly statistically significant difference from the T-test.

## Conclusion

From these observations we have highlighted the use of Fmax_mean as an alternative diagnostic variable both for the tests done 24 and 150 hours after the sampling. Thus we can qualify both methods as viable testing methods for diagnostics. The data set includes 4 other variables that may also be suited for this(AUC, SLOPE, T50, TTT), the above process could simply be repeated for each one in order to evaluate them.
